+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/_directory_client.py
@@ -30,6 +30,7 @@ from ._shared.response_handlers import return_response_headers, process_storage_
 from ._shared.parser import _str
 from ._parser import _get_file_permission, _datetime_to_str
 from ._deserialize import deserialize_directory_properties
+from ._serialize import get_api_version
 from ._file_client import ShareFileClient
 from ._models import DirectoryPropertiesPaged, HandlesPaged, NTFSAttributes  # pylint: disable=unused-import
 
@@ -61,6 +62,12 @@ class ShareDirectoryClient(StorageAccountHostsMixin):
         The credential with which to authenticate. This is optional if the
         account URL already has a SAS token. The value can be a SAS token string or an account
         shared access key.
+    :keyword str api_version:
+        The Storage API version to use for requests. Default value is '2019-07-07'.
+        Setting to an older version may result in reduced feature compatibility.
+
+        .. versionadded:: 12.1.0
+
     :keyword str secondary_hostname:
         The hostname of the secondary endpoint.
     :keyword int max_range_size: The maximum range size used for a file upload. Defaults to 4*1024*1024.
@@ -106,6 +113,7 @@ class ShareDirectoryClient(StorageAccountHostsMixin):
             sas_token, credential, share_snapshot=self.snapshot)
         super(ShareDirectoryClient, self).__init__(parsed_url, service='file-share', credential=credential, **kwargs)
         self._client = AzureFileStorage(version=VERSION, url=self.url, pipeline=self._pipeline)
+        self._client._config.version = get_api_version(kwargs, VERSION)  # pylint: disable=protected-access
 
     @classmethod
     def from_directory_url(cls, directory_url,  # type: str
@@ -216,7 +224,8 @@ class ShareDirectoryClient(StorageAccountHostsMixin):
         )
         return ShareFileClient(
             self.url, file_path=file_name, share_name=self.share_name, napshot=self.snapshot,
-            credential=self.credential, _hosts=self._hosts, _configuration=self._config,
+            credential=self.credential, api_version=self.api_version,
+            _hosts=self._hosts, _configuration=self._config,
             _pipeline=_pipeline, _location_mode=self._location_mode, **kwargs)
 
     def get_subdirectory_client(self, directory_name, **kwargs):
@@ -247,7 +256,8 @@ class ShareDirectoryClient(StorageAccountHostsMixin):
         )
         return ShareDirectoryClient(
             self.url, share_name=self.share_name, directory_path=directory_path, snapshot=self.snapshot,
-            credential=self.credential, _hosts=self._hosts, _configuration=self._config, _pipeline=_pipeline,
+            credential=self.credential, api_version=self.api_version,
+            _hosts=self._hosts, _configuration=self._config, _pipeline=_pipeline,
             _location_mode=self._location_mode, **kwargs)
 
     @distributed_trace
@@ -399,6 +409,7 @@ class ShareDirectoryClient(StorageAccountHostsMixin):
             )
             return {
                 'closed_handles_count': response.get('number_of_handles_closed', 0),
+                'failed_handles_count': response.get('number_of_handles_failed', 0)
             }
         except StorageErrorException as error:
             process_storage_error(error)
@@ -425,6 +436,7 @@ class ShareDirectoryClient(StorageAccountHostsMixin):
         try_close = True
         continuation_token = None
         total_closed = 0
+        total_failed = 0
         while try_close:
             try:
                 response = self._client.directory.force_close_handles(
@@ -441,10 +453,12 @@ class ShareDirectoryClient(StorageAccountHostsMixin):
             continuation_token = response.get('marker')
             try_close = bool(continuation_token)
             total_closed += response.get('number_of_handles_closed', 0)
+            total_failed += response.get('number_of_handles_failed', 0)
             if timeout:
                 timeout = max(0, timeout - (time.time() - start_time))
         return {
             'closed_handles_count': total_closed,
+            'failed_handles_count': total_failed
         }
 
     @distributed_trace