+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/aio/__init__.py
@@ -8,6 +8,7 @@ from ._file_client_async import ShareFileClient
 from ._directory_client_async import ShareDirectoryClient
 from ._share_client_async import ShareClient
 from ._share_service_client_async import ShareServiceClient
+from ._lease_async import ShareLeaseClient
 
 
 __all__ = [
@@ -15,4 +16,5 @@ __all__ = [
     'ShareDirectoryClient',
     'ShareClient',
     'ShareServiceClient',
+    'ShareLeaseClient',
 ]