+++ sdk/storage/azure-storage-file-share/tests/_shared/asynctestcase.py
@@ -12,7 +12,7 @@ from azure_devtools.scenario_tests.utilities import trim_kwargs_from_test_functi
 
 from azure.core.credentials import AccessToken
 
-from .filetestcase import FileTestCase
+from .testcase import StorageTestCase
 
 LOGGING_FORMAT = '%(asctime)s %(name)-20s %(levelname)-5s %(message)s'
 
@@ -27,7 +27,7 @@ class AsyncFakeTokenCredential(object):
         return self.token
 
 
-class AsyncStorageTestCase(FileTestCase):
+class AsyncStorageTestCase(StorageTestCase):
     @staticmethod
     def await_prepared_test(test_fn):
         """Synchronous wrapper for async test methods. Used to avoid making changes